requirements.txt
google-generativeai>=0.8.3
python-dotenv>=1.0.1
pydantic>=2.7
requests>=2.31.0
moviepy==1.0.3
Pillow>=10.3.0
numpy>=1.26
tqdm>=4.66
pydub>=0.25.1
# Optional local transcription if needed later
faster-whisper>=1.0.0
imageio-ffmpeg>=0.4.9